VERSION 1.0

+++ lib/Screens/home/components/DietListView.dart
@@ -21,7 +21,7 @@ Future getData() async{
   for (var i=0;i<listPrefData.length;i++) {
     var name = listPrefData[i];
     List<String> listPrefData2 = pref.getStringList(name);
-    FoodModel foodModel = FoodModel(name: name, calo100g: double.parse(listPrefData2[0]), quantity: int.parse(listPrefData2[1]));
+    FoodModel foodModel = FoodModel(name: name, calo100g: double.parse(listPrefData2[0]), quantity: double.parse(listPrefData2[1]));
     listData.add(foodModel);
   }
 }
